+++ src/jp_learning_platform/application/__init__.py
@@ -1,3 +1,25 @@
 """Application use-case layer."""
 
-__all__: list[str] = []
+from jp_learning_platform.application.subtitle_pipeline import (
+    AudioInputDiscovery,
+    DEFAULT_OUTPUT_DIRECTORY,
+    InputPathNotFoundError,
+    NoAudioInputsFoundError,
+    SUPPORTED_AUDIO_EXTENSIONS,
+    SubtitlePipelineInputError,
+    SubtitlePipelineItemResult,
+    SubtitlePipelineRequest,
+    SubtitlePipelineResult,
+)
+
+__all__ = [
+    "AudioInputDiscovery",
+    "DEFAULT_OUTPUT_DIRECTORY",
+    "InputPathNotFoundError",
+    "NoAudioInputsFoundError",
+    "SUPPORTED_AUDIO_EXTENSIONS",
+    "SubtitlePipelineInputError",
+    "SubtitlePipelineItemResult",
+    "SubtitlePipelineRequest",
+    "SubtitlePipelineResult",
+]